_helper/uswb/requirements.txt
crc
pyserial
